Added isser to get if a day is considered a holiday or not and changed folder structure.
