+++ Date/NewYearsEve.php
@@ -4,18 +4,14 @@ declare (strict_types=1);
 
 namespace JGI\SwedishDates\Date;
 
-class NationalDay implements DayInterface, DayOccurOnceAYearInterface
+class NewYearsEve implements DayInterface, DayOccurOnceAYearInterface
 {
     /**
      * {@inheritdoc}
      */
     public function match(\DateTime $datetime): bool
     {
-        if ($nationalDay = $this->getDateForYear((int) $datetime->format('Y'))) {
-            return $nationalDay->format('Y-m-d') == $datetime->format('Y-m-d');
-        }
-
-        return false;
+        return $this->getDateForYear((int) $datetime->format('Y'))->format('Y-m-d') == $datetime->format('Y-m-d');
     }
 
     /**
@@ -29,9 +25,17 @@ class NationalDay implements DayInterface, DayOccurOnceAYearInterface
     /**
      * {@inheritdoc}
      */
+    public function isHoliday(): bool
+    {
+        return true;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function getName(): string
     {
-        return 'Nationaldagen';
+        return 'Nyårsafton';
     }
 
     /**
@@ -39,10 +43,6 @@ class NationalDay implements DayInterface, DayOccurOnceAYearInterface
      */
     public function getDateForYear(int $year)
     {
-        if ($year >= 2005) {
-            return new \DateTime(sprintf('%s-06-06', $year));
-        }
-
-        return null;
+        return new \DateTime(sprintf('%s-12-31', $year));
     }
 }
