+++ Tests/Date/AllSaintsDayTest.php
