+++ Tests/Date/MidsummerEveTest.php
