+++ composer.json
@@ -12,9 +12,12 @@
     "require": {
         "php": ">=7.0"
     },
+    "require-dev": {
+        "phpunit/phpunit": "^6.0"
+    },
     "autoload": {
         "psr-4": {
-            "JGI\\SwedishDates\\": "src/JGI/SwedishDates/"
+            "JGI\\SwedishDates\\": ""
         }
     }
 }
