+++ phpunit.xml.dist
@@ -9,11 +9,11 @@
          processIsolation="false"
          stopOnFailure="false"
          syntaxCheck="false"
-         bootstrap="tests/bootstrap.php"
+         bootstrap="vendor/autoload.php"
 >
     <testsuites>
         <testsuite name="SwedishDates Test Suite">
-            <directory>./tests/JGI/SwedishDates/</directory>
+            <directory>Tests</directory>
         </testsuite>
     </testsuites>
