--- src/JGI/SwedishDates/Date/AscensionDay.php
@@ -1,44 +1,0 @@
-<?php
-
-declare (strict_types=1);
-
-namespace JGI\SwedishDates\Date;
-
-class AscensionDay implements DayInterface, DayOccurOnceAYearInterface
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function match(\DateTime $datetime): bool
-    {
-        return $this->getDateForYear((int) $datetime->format('Y'))->format('Y-m-d') == $datetime->format('Y-m-d');
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isRed(): bool
-    {
-        return true;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName(): string
-    {
-        return 'Kristi Himmelsfärdsdag';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDateForYear(int $year)
-    {
-        $datetime = new \DateTime();
-        $datetime->setTimestamp(easter_date($year));
-        $datetime->modify('+39 day');
-
-        return $datetime;
-    }
-}
